CTCP-534: Amending test-only routes and adding method for sending departure message.
