+++ app/connectors/testOnly/TestOnlyRouterConnector.scala
@@ -19,7 +19,7 @@ package connectors.testOnly
 import config.FrontendAppConfig
 import play.api.mvc.Headers
 import uk.gov.hmrc.http.HttpReads.Implicits._
-import uk.gov.hmrc.http.{Authorization, HeaderCarrier, HttpClient, HttpReads, HttpResponse}
+import uk.gov.hmrc.http.{HeaderCarrier, HttpClient, HttpResponse}
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
@@ -27,51 +27,19 @@ import scala.xml.NodeSeq
 
 class TestOnlyRouterConnector @Inject() (val http: HttpClient, config: FrontendAppConfig)(implicit ec: ExecutionContext) {
 
-  def submitInboundMessage(requestData: NodeSeq, headers: Headers)(implicit headerCarrier: HeaderCarrier): Future[HttpResponse] = {
-
+  def submitInboundMessage(
+    requestData: NodeSeq,
+    headers: Headers
+  )(implicit headerCarrier: HeaderCarrier): Future[HttpResponse] = {
     val routerUrl = s"${config.routerUrl}/messages"
 
     val header = headers.headers.filter(
-      x => x._1.equalsIgnoreCase("X-Message-Recipient") || x._1.equalsIgnoreCase("X-Message-Type") || x._1.equalsIgnoreCase("Content-Type")
+      x =>
+        x._1.equalsIgnoreCase("X-Message-Recipient") ||
+          x._1.equalsIgnoreCase("X-Message-Type") ||
+          x._1.equalsIgnoreCase("Content-Type")
     )
 
     http.POSTString[HttpResponse](routerUrl, requestData.toString, header)
-  }
-
-  private def addHeaders(): Seq[(String, String)] = Seq("Content-Type" -> "application/xml", "Channel" -> "web")
-
-  def createArrivalNotificationMessage(requestData: NodeSeq, headers: Headers)(implicit headerCarrier: HeaderCarrier): Future[HttpResponse] = {
-
-    val serviceUrl = s"${config.destinationUrl}/movements/arrivals"
-
-    val newHeaders = headerCarrier
-      .copy(authorization = Some(Authorization(headers.get("Authorization").getOrElse(""))))
-      .withExtraHeaders(addHeaders(): _*)
-
-    http.POSTString[HttpResponse](serviceUrl, requestData.toString)(rds = HttpReads[HttpResponse], hc = newHeaders, ec = ec)
-  }
-
-  def resubmitArrivalNotificationMessage(requestData: NodeSeq, arrivalId: String, headers: Headers)(implicit
-    headerCarrier: HeaderCarrier
-  ): Future[HttpResponse] = {
-
-    val serviceUrl = s"${config.destinationUrl}/movements/arrivals/$arrivalId"
-
-    val newHeaders = headerCarrier
-      .copy(authorization = Some(Authorization(headers.get("Authorization").getOrElse(""))))
-      .withExtraHeaders(addHeaders(): _*)
-
-    http.PUTString[HttpResponse](serviceUrl, requestData.toString)(rds = HttpReads[HttpResponse], hc = newHeaders, ec = ec)
-  }
-
-  def submitMessageToCore(requestData: NodeSeq, arrivalId: String, headers: Headers)(implicit headerCarrier: HeaderCarrier): Future[HttpResponse] = {
-
-    val serviceUrl = s"${config.destinationUrl}/movements/arrivals/$arrivalId/messages"
-
-    val newHeaders = headerCarrier
-      .copy(authorization = Some(Authorization(headers.get("Authorization").getOrElse(""))))
-      .withExtraHeaders(addHeaders(): _*)
-
-    http.POSTString[HttpResponse](serviceUrl, requestData.toString)(rds = HttpReads[HttpResponse], hc = newHeaders, ec = ec)
   }
 }
