+++ app/controllers/testOnly/TestOnlyRouterController.scala
@@ -21,7 +21,7 @@ import play.api.mvc.{Action, DefaultActionBuilder, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 
 import javax.inject.Inject
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 import scala.xml.NodeSeq
 
 class TestOnlyRouterController @Inject() (
@@ -38,54 +38,5 @@ class TestOnlyRouterController @Inject() (
         .map(
           response => Status(response.status)
         )
-  }
-
-  def arrivalNotificationMessageToCore: Action[NodeSeq] = action.async(parse.xml) {
-    implicit request =>
-      connector
-        .createArrivalNotificationMessage(request.body, request.headers)
-        .map {
-          response =>
-            val location = response.header("Location").getOrElse("Location is missing")
-            Status(response.status)
-              .withHeaders(
-                "Location"  -> location,
-                "arrivalId" -> location.split("/").last
-              )
-        }
-  }
-
-  def resubmitArrivalNotificationMessageToCore: Action[NodeSeq] = action.async(parse.xml) {
-    implicit request =>
-      request.headers.get("arrivalId") match {
-        case Some(arrivalId) =>
-          connector
-            .resubmitArrivalNotificationMessage(request.body, arrivalId, request.headers)
-            .map {
-              response =>
-                val location = response.header("Location").getOrElse("Location is missing")
-                Status(response.status)
-                  .withHeaders(
-                    "Location"  -> location,
-                    "arrivalId" -> location.split("/").last
-                  )
-            }
-
-        case _ => Future.successful(BadRequest("ArrivalId is missing"))
-      }
-  }
-
-  def messageToCore: Action[NodeSeq] = action.async(parse.xml) {
-    implicit request =>
-      request.headers.get("arrivalId") match {
-        case Some(arrivalId) =>
-          connector
-            .submitMessageToCore(request.body, arrivalId, request.headers)
-            .map(
-              response => Status(response.status)
-            )
-
-        case _ => Future.successful(BadRequest("ArrivalId is missing"))
-      }
   }
 }
