+++ conf/application.conf
@@ -84,6 +84,13 @@ microservice {
       startUrl = "transit-movements-trader-at-destination"
     }
 
+    test-support {
+      protocol = http
+      host = localhost
+      port = 9497
+      startUrl = ""
+    }
+
 #  Warning: This must remain false in main, these will be toggled by service manager profiles
 # otherwise users in production will be pointed to new phase 5 frontends
     features {
